Added new tests and scope handling to antipattern
Added scope handling to the antipattern, and a few new test files so we can check this. I made a separate folder in cpp_tests/ for my tests, since I also plan on adding testing for do-while loops / for loops etc..
And for lock_guarded function calls (similar to calculate in order.cpp)

Pytest will take longer to run with these, but this is because there are 2 more files being ran. When using clang to go through nodes in a file, it will also go through all the nodes in the #include statements recursively which is what mainly slows the process down. Only nodes inside the actual cpp file we want to check will be fed into the checks (none of the includes will be fed into the checks).