Antipattern done - Added lock/unlock handling
This antipattern works fully now (i think). The only thing left for me to do is make it so that I don't need to walk_preorder inside the python functions. I could do this now, but it might be wiser to prioritize the next antipatterns and then come back to that if I have time at the end?
I still need to add tests for the lock/unlocks.