Finished antipattern, added tests for locks/unlocks
Refactored antipattern code so that it will only do walk_preorder() once.
The only time nodes are ran through twice now is when we are looking for the data members of the class using get_children() (which is necessary)

Added tests which include manual locks/unlocks. Including a combination of manual locks/unlocks and lock_guards