Updated antipattern to reflect new traversal methods in main
Updated antipattern to reflect new traversal methods

Added tests to reflect this
Also adjusted main so that it would not see observer.py as a check, by adding it to the if statement on line 39.

There is an issue with parsing idx.parse with an additional arg with the examples we were provided, so that arg is removed in this branch

The tests for observer aren't checks so running the run_check_on_file function wouldn't suit it. Instead, I added my searchNodes function to observer.py solely to have a function to test observer.py with. It is only for tests, and is not needed in the actual program.