Added code to handle if statements
Added code to handle the condition of "If we enter an if-else body and change the value of an atomic whose value affected the condition"

This is still very much a work in progress, nested if statements are yet to be tested.

I don't have the locked-calling atomic code in this as I wanted a blank canvas to work from, I plan on integrating it into locked-calling in the very near future.