Testing with nested branches, start of set up of other statements
Started integrated nested branches into test.cpp files for atomic, they work with the limited testing done so far.

Started the handling for statements that have a condition and body, not just if statements.

Currently thinking of using an object for each statement, and removing the object once we've ran through all the cursors in the statement. These objects will be stored in a dict() and replace atomic_if_conditions.