Switched to use object instead of cursor, added tests
Currently facing a problem with else statements. I know how to fix it and am working through it right now.

The issue is that I am only accounting for one compound statement in the object, when i need to account for multiple (for if-else etc..), I can fix it using a list of compounds instead.